refactor(framework): Move reflection code into ReflectionSidecar
