+++ framework/src/main/java/org/fulib/fx/controller/internal/ReflectionSidecar.java
@@ -1,0 +1,455 @@
+package org.fulib.fx.controller.internal;
+
+import javafx.beans.value.WritableValue;
+import javafx.scene.Node;
+import javafx.scene.Parent;
+import org.fulib.fx.FulibFxApp;
+import org.fulib.fx.annotation.controller.*;
+import org.fulib.fx.annotation.event.onDestroy;
+import org.fulib.fx.annotation.event.onInit;
+import org.fulib.fx.annotation.event.onRender;
+import org.fulib.fx.annotation.param.Param;
+import org.fulib.fx.annotation.param.Params;
+import org.fulib.fx.annotation.param.ParamsMap;
+import org.fulib.fx.controller.ControllerManager;
+import org.fulib.fx.util.ControllerUtil;
+import org.fulib.fx.util.MapUtil;
+import org.fulib.fx.util.ReflectionUtil;
+import org.fulib.fx.util.reflection.Reflection;
+import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
+import org.jetbrains.annotations.Unmodifiable;
+
+import java.lang.annotation.Annotation;
+import java.lang.reflect.Field;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.lang.reflect.Modifier;
+import java.util.*;
+
+import static org.fulib.fx.util.FrameworkUtil.error;
+
+public class ReflectionSidecar<T> implements FxSidecar<T> {
+    private final ControllerManager controllerManager;
+
+    public ReflectionSidecar(ControllerManager controllerManager) {
+        this.controllerManager = controllerManager;
+    }
+
+    @Override
+    public void init(T instance, Map<String, Object> params) {
+        // Inject parameters into the controller fields
+        fillParametersIntoFields(instance, params);
+
+        // Call parameter setter methods
+        callParamMethods(instance, params);
+        callParamsMethods(instance, params);
+        callParamsMapMethods(instance, params);
+
+        // Call the onInit method(s)
+        callMethodsWithAnnotation(instance, onInit.class, params);
+
+        // Initialize all subcomponents
+        Reflection.callMethodsForFieldInstances(instance, getSubComponentFields(instance), (subController) -> controllerManager.init(subController, params));
+    }
+
+    /**
+     * Calls all methods annotated with a certain annotation in the provided controller. The method will be called with the given parameters if they're annotated with @Param or @Params.
+     *
+     * @param annotation The annotation to look for
+     * @param parameters The parameters to pass to the methods
+     */
+    private void callMethodsWithAnnotation(@NotNull Object instance, @NotNull Class<? extends Annotation> annotation, @NotNull Map<@NotNull String, @Nullable Object> parameters) {
+        for (Method method : Reflection.getAllMethodsWithAnnotation(instance.getClass(), annotation).sorted(annotationComparator(annotation)).toList()) {
+            try {
+                method.setAccessible(true);
+                method.invoke(instance, getApplicableParameters(method, parameters));
+            } catch (IllegalAccessException | InvocationTargetException e) {
+                throw new RuntimeException(error(1005).formatted(method.getName(), annotation.getName(), instance.getClass().getName()), e);
+            }
+        }
+    }
+
+    /**
+     * Returns a comparator that compares methods based on the value of the given annotation.
+     * The method assumes that the annotation has a method called 'value' that returns an integer value.
+     *
+     * @param annotation The annotation to compare
+     * @return A comparator that compares methods based on the value of the given annotation
+     */
+    private static Comparator<Method> annotationComparator(@NotNull Class<? extends Annotation> annotation) {
+        return (m1, m2) -> {
+            Annotation event1 = m1.getAnnotation(annotation);
+            Annotation event2 = m2.getAnnotation(annotation);
+            try {
+                Method value = annotation.getDeclaredMethod("value");
+                return Integer.compare((int) value.invoke(event1), (int) value.invoke(event2));
+            } catch (ReflectiveOperationException e) {
+                return 0;
+            }
+        };
+    }
+
+    /**
+     * Fills the parameters map into the fields annotated with @Param and @ParamsMap in the given instance.
+     * (Parameter/Map injection into fields)
+     *
+     * @param instance   The instance to fill the parameters into
+     * @param parameters The parameters to fill into the fields
+     */
+    private void fillParametersIntoFields(@NotNull Object instance, @NotNull Map<@NotNull String, @Nullable Object> parameters) {
+        // Fill the parameters into fields annotated with @Param
+        for (Field field : Reflection.getAllFieldsWithAnnotation(instance.getClass(), Param.class).toList()) {
+
+            Param paramAnnotation = field.getAnnotation(Param.class);
+            String param = paramAnnotation.value();
+
+            // Don't fill the parameter if it's not present (field will not be overwritten, "default value")
+            if (!parameters.containsKey(param)) {
+                continue;
+            }
+
+            Class<?> fieldType = field.getType();
+            try {
+                field.setAccessible(true);
+
+                Object value = parameters.get(param);
+                Object fieldValue = field.get(instance);
+
+                // If the field is a WriteableValue, use the setValue method
+                if (WritableValue.class.isAssignableFrom(fieldType) && !(value instanceof WritableValue)) {
+
+                    // We cannot call setValue on a non-existing property
+                    if (fieldValue == null) {
+                        throw new RuntimeException(error(4001).formatted(param, field.getName(), instance.getClass().getName()));
+                    }
+
+                    try {
+                        // noinspection unchecked
+                        ((WritableValue<Object>) field.get(instance)).setValue(value);
+                    } catch (ClassCastException e) {
+                        throw new RuntimeException(error(4007).formatted(param, field.getName(), instance.getClass().getName(), fieldType.getName(), value == null ? "null" : value.getClass().getName()));
+                    }
+                }
+
+                // If not, set the field's value directly
+                else if (value == null) {
+                    // If the value is null and the field is a primitive, throw an error
+                    if (fieldType.isPrimitive()) {
+                        throw new RuntimeException(error(4007).formatted(param, field.getName(), instance.getClass().getName(), fieldType.getName(), "null"));
+                    }
+                    field.set(instance, null); // If the value is null and the field is not a primitive, no type check is necessary
+                } else if (Reflection.canBeAssigned(fieldType, value)) {
+                    field.set(instance, value); // If the value is not null, we need a type check (respects primitive types)
+                } else {
+                    throw new RuntimeException(error(4007).formatted(param, field.getName(), instance.getClass().getName(), fieldType.getName(), value.getClass().getName()));
+                }
+
+            } catch (IllegalAccessException e) {
+                throw new RuntimeException(error(4000).formatted(param, field.getName(), instance.getClass().getName()), e);
+            }
+        }
+
+        // Fill the parameters into fields annotated with @ParamsMap
+        for (Field field : Reflection.getFieldsWithAnnotation(instance.getClass(), ParamsMap.class).toList()) {
+
+            if (!MapUtil.isMapWithTypes(field, String.class, Object.class)) {
+                throw new RuntimeException(error(4002).formatted(field.getName(), instance.getClass().getName()));
+            }
+
+            try {
+                field.setAccessible(true);
+
+                // If the map is final, clear it and put all parameters into it
+                if (Modifier.isFinal(field.getModifiers())) {
+                    @SuppressWarnings("unchecked")
+                    Map<String, Object> map = (Map<String, Object>) field.get(instance);
+                    map.clear();
+                    map.putAll(parameters);
+                } else {
+                    field.set(instance, parameters);
+                }
+            } catch (IllegalAccessException e) {
+                throw new RuntimeException(error(4010).formatted(field.getName(), instance.getClass().getName()), e);
+            }
+        }
+    }
+
+    /**
+     * Calls all methods annotated with @Param in the given instance with the values of the specified parameter.
+     * (Parameter injection into setters)
+     *
+     * @param instance   The instance to fill the parameters into
+     * @param parameters The parameters to fill into the methods
+     */
+    private void callParamMethods(Object instance, Map<String, Object> parameters) {
+        Reflection.getAllMethodsWithAnnotation(instance.getClass(), Param.class).forEach(method -> {
+            try {
+                method.setAccessible(true);
+                Object value = parameters.get(method.getAnnotation(Param.class).value());
+
+                if (value == null) {
+                    method.invoke(instance, (Object) null);
+                    return;
+                }
+
+                if (Reflection.canBeAssigned(method.getParameterTypes()[0], value)) {
+                    method.invoke(instance, value);
+                } else {
+                    throw new RuntimeException(error(4008).formatted(method.getAnnotation(Param.class).value(), method.getName(), instance.getClass().getName(), method.getParameterTypes()[0].getName(), value.getClass().getName()));
+                }
+            } catch (IllegalAccessException | InvocationTargetException e) {
+                throw new RuntimeException(error(4005).formatted(method.getAnnotation(Param.class).value(), method.getName(), instance.getClass().getName()), e);
+            }
+        });
+    }
+
+    /**
+     * Calls all methods annotated with @Params in the given instance with the values of the specified parameters.
+     * (Multiple parameter injection into methods)
+     *
+     * @param instance   The instance to fill the parameters into
+     * @param parameters The parameters to fill into the methods
+     */
+    private void callParamsMethods(Object instance, Map<String, Object> parameters) {
+        Reflection.getAllMethodsWithAnnotation(instance.getClass(), Params.class).forEach(method -> {
+            try {
+                method.setAccessible(true);
+
+                String[] paramNames = method.getAnnotation(Params.class).value();
+                if (method.getParameters().length != paramNames.length) {
+                    throw new RuntimeException(error(4006).formatted(method.getName(), instance.getClass().getName()));
+                }
+
+                Object[] methodParams = new Object[paramNames.length];
+
+                // Fill the parameters into the method
+                for (int i = 0; i < paramNames.length; i++) {
+                    Object value = parameters.get(paramNames[i]);
+                    if (Reflection.canBeAssigned(method.getParameterTypes()[i], value)) {
+                        methodParams[i] = value;
+                    } else {
+                        throw new RuntimeException(error(4008).formatted(paramNames[i], method.getName(), instance.getClass().getName(), method.getParameterTypes()[i].getName(), value.getClass().getName()));
+                    }
+                }
+
+                method.invoke(instance, methodParams);
+            } catch (IllegalAccessException | InvocationTargetException e) {
+                throw new RuntimeException(error(4011).formatted(method.getName(), instance.getClass().getName()), e);
+            }
+        });
+    }
+
+    /**
+     * Fills the parameters map into the methods annotated with @ParamsMap in the given instance.
+     * (Map injection into setters)
+     *
+     * @param instance   The instance to fill the parameters into
+     * @param parameters The parameters to fill into the methods
+     */
+    private void callParamsMapMethods(Object instance, Map<String, Object> parameters) {
+        Reflection.getAllMethodsWithAnnotation(instance.getClass(), ParamsMap.class).forEach(method -> {
+
+            if (method.getParameterCount() != 1 || !MapUtil.isMapWithTypes(method.getParameters()[0], String.class, Object.class)) {
+                throw new RuntimeException(error(4003).formatted(method.getName(), instance.getClass().getName()));
+            }
+
+            try {
+                method.setAccessible(true);
+                method.invoke(instance, parameters);
+            } catch (IllegalAccessException | InvocationTargetException e) {
+                throw new RuntimeException(error(4010).formatted(method.getName(), instance.getClass().getName()), e);
+            }
+        });
+    }
+
+    /**
+     * Returns an array with all parameters that are applicable to the given method in the correct order.
+     * <p>
+     * If the method has a parameter annotated with @Param, the value of the parameter with the same key as the annotation will be used.
+     * <p>
+     * If the method has a parameter annotated with @ParamsMap, the whole parameters map will be used.
+     *
+     * @param method     The method to check
+     * @param parameters The values of the parameters
+     * @return An array with all applicable parameters
+     */
+    private @Nullable Object @NotNull [] getApplicableParameters(@NotNull Method method, @NotNull Map<String, Object> parameters) {
+        return Arrays.stream(method.getParameters()).map(parameter -> {
+            Param param = parameter.getAnnotation(Param.class);
+            ParamsMap paramsMap = parameter.getAnnotation(ParamsMap.class);
+
+            if (param != null && paramsMap != null) {
+                throw new RuntimeException(error(4009).formatted(parameter.getName(), method.getName(), method.getDeclaringClass().getName()));
+            }
+
+            // Check if the parameter is annotated with @Param and if the parameter is of the correct type
+            if (param != null) {
+                if (parameters.containsKey(param.value()) && !Reflection.canBeAssigned(parameter.getType(), parameters.get(param.value()))) {
+                    throw new RuntimeException(error(4008).formatted(param.value(), method.getName(), method.getDeclaringClass().getName(), parameter.getType().getName(), parameters.get(param.value()).getClass().getName()));
+                }
+                return parameters.get(param.value());
+            }
+
+            // Check if the parameter is annotated with @Params and if the parameter is of the type Map<String, Object>
+            if (paramsMap != null) {
+                if (!MapUtil.isMapWithTypes(parameter, String.class, Object.class)) {
+                    throw new RuntimeException(error(4004).formatted(parameter.getName(), method.getName(), method.getDeclaringClass().getName()));
+                }
+                return parameters;
+            }
+            return null;
+        }).toArray();
+    }
+
+
+    /**
+     * Returns a list of all fields in the given instance that are annotated with {@link SubComponent}.
+     *
+     * @param instance The instance to get the fields from
+     * @return A list of all fields in the given instance that are annotated with {@link SubComponent}
+     */
+    @Unmodifiable
+    private List<Field> getSubComponentFields(Object instance) {
+        return Reflection.getAllFieldsWithAnnotation(instance.getClass(), SubComponent.class)
+            .filter(field -> {
+                if (ControllerUtil.isComponent(field.getType())) {
+                    return true;
+                }
+
+                if (!ControllerUtil.canProvideSubComponent(field)) {
+                    FulibFxApp.LOGGER.warning(error(6005).formatted(field.getName(), instance.getClass().getName()));
+                }
+                return false;
+            }).toList();
+    }
+
+    @Override
+    public Node render(T instance, Map<String, Object> params) {
+        // Render all subcomponents
+        Reflection.callMethodsForFieldInstances(instance, getSubComponentFields(instance), (subController) -> controllerManager.render(subController, params));
+
+        // Get the view of the controller
+        final boolean component = ControllerUtil.isComponent(instance);
+        final Node node = renderNode(instance, component);
+
+        // Call the onRender method
+        callMethodsWithAnnotation(instance, onRender.class, params);
+
+        return node;
+    }
+
+    private Node renderNode(T instance, boolean component) {
+        Node node;
+        String view = component ?
+            instance.getClass().getAnnotation(Component.class).view() :
+            instance.getClass().getAnnotation(Controller.class).view();
+
+        // If the controller extends from a javafx Node, render it
+        // This can be combined with the view annotation to set the controller as the root of the fxml file
+        if (component) {
+            if (view.isEmpty()) {
+                node = (Node) instance;
+            } else {
+                Node root = (Node) instance;
+                // Due to the way JavaFX works, we have to clear the children list of the old root before loading its fxml file again
+                if (root instanceof Parent parent) {
+                    ReflectionUtil.getChildrenList(instance.getClass(), parent).clear();
+                }
+                node = controllerManager.loadFXML(view, instance, true);
+            }
+        }
+
+        // If the controller specifies a method returning a parent as its view, call it
+        else if (view.startsWith("#")) {
+            String methodName = view.substring(1);
+            try {
+                Method method = instance.getClass().getDeclaredMethod(methodName);
+                if (method.getParameterCount() != 0) {
+                    throw new RuntimeException(error(1008).formatted(methodName, instance.getClass().getName()));
+                }
+                if (!Parent.class.isAssignableFrom(method.getReturnType())) {
+                    throw new RuntimeException(error(1002).formatted(methodName, instance.getClass().getName()));
+                }
+                method.setAccessible(true);
+                node = (Parent) method.invoke(instance);
+            } catch (NoSuchMethodException e) {
+                throw new RuntimeException(error(1003).formatted(methodName, instance.getClass().getName()), e);
+            } catch (InvocationTargetException | IllegalAccessException e) {
+                throw new RuntimeException(error(1004).formatted(methodName, instance.getClass().getName()), e);
+            }
+        }
+
+        // If the controller specifies a fxml file, load it. This will also load subcomponents specified in the FXML file
+        else {
+            String fxmlPath = view.isEmpty() ? ControllerUtil.transform(instance.getClass().getSimpleName()) + ".fxml" : view;
+            node = controllerManager.loadFXML(fxmlPath, instance, false);
+        }
+        return node;
+    }
+
+    @Override
+    public void destroy(T instance) {
+        // Destroying should be done in exactly the reverse order of initialization
+        List<Field> subComponentFields = new ArrayList<>(getSubComponentFields(instance));
+        Collections.reverse(subComponentFields);
+
+        // Destroy all subcomponents
+        Reflection.callMethodsForFieldInstances(instance, subComponentFields, controllerManager::destroy);
+
+        // Call destroy methods
+        callMethodsWithAnnotation(instance, onDestroy.class, Map.of());
+    }
+
+    @Override
+    public @Nullable ResourceBundle getResources(T instance) {
+        List<Field> fields = Reflection.getAllFieldsWithAnnotation(instance.getClass(), Resource.class).toList();
+
+        if (fields.isEmpty()) {
+            return controllerManager.getDefaultResourceBundle();
+        }
+
+        if (fields.size() > 1) {
+            throw new RuntimeException(error(2003).formatted(instance.getClass().getName()));
+        }
+
+        return fields
+            .stream()
+            .filter(field -> {
+                if (field.getType().isAssignableFrom(ResourceBundle.class)) {
+                    return true;
+                }
+                throw new RuntimeException(error(2004).formatted(field.getName(), instance.getClass().getName()));
+            })
+            .map(field -> {
+                try {
+                    field.setAccessible(true);
+                    return (ResourceBundle) field.get(instance);
+                } catch (IllegalAccessException e) {
+                    throw new RuntimeException(error(2005).formatted(field.getName(), instance.getClass().getName()), e);
+                }
+            })
+            .filter(Objects::nonNull)
+            .findFirst()
+            .orElse(controllerManager.getDefaultResourceBundle());
+    }
+
+    @Override
+    public @Nullable String getTitle(T instance) {
+        if (!instance.getClass().isAnnotationPresent(Title.class)) {
+            return null;
+        }
+
+        String title = instance.getClass().getAnnotation(Title.class).value();
+
+        if (title.startsWith("%")) {
+            title = title.substring(1);
+            ResourceBundle resourceBundle = getResources(instance);
+            return resourceBundle.getString(title);
+        } else if ("$name".equals(title)) {
+            return ControllerUtil.transform(instance.getClass().getSimpleName());
+        }
+
+        return title;
+    }
+}
